+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Improve `RumUserActionDetector` to detect more widgets, including `BottomNavigationBar`, `Tab`, `Switch`, and `Radio`
 * Remove an extra call to `FlutterError.presentError` made in `runApp`.  See [#358]
 * Set `sessionReplaySampleRate` to 0 during initialization for Browser as Session Replay is not supported.
+* Support `errorType` on `DdRum.addError` and `DdRum.addErrorInfo`. See [#372]
 
 ## 1.2.2
 
@@ -150,3 +151,4 @@
 [#328]: https://github.com/DataDog/dd-sdk-flutter/issues/328
 [#334]: https://github.com/DataDog/dd-sdk-flutter/issues/334
 [#358]: https://github.com/DataDog/dd-sdk-flutter/issues/358
+[#372]: https://github.com/DataDog/dd-sdk-flutter/issues/372